+++ internal/fileio/preallocate_darwin.go
@@ -1,4 +1,4 @@
-package fs
+package fileio
 
 import (
 	"os"
