+++ internal/fileio/preallocate_linux.go
@@ -1,4 +1,4 @@
-package fs
+package fileio
 
 import (
 	"os"
