+++ internal/fileio/preallocate_other.go
@@ -1,6 +1,6 @@
 //go:build !linux && !darwin
 
-package fs
+package fileio
 
 import "os"
