+++ internal/fileio/preallocate_test.go
@@ -1,4 +1,4 @@
-package fs
+package fileio
 
 import (
 	"os"
@@ -7,6 +7,7 @@ import (
 	"syscall"
 	"testing"
 
+	"github.com/restic/restic/internal/fs"
 	"github.com/restic/restic/internal/test"
 )
 
@@ -31,7 +32,7 @@ func TestPreallocate(t *testing.T) {
 			fi, err := wr.Stat()
 			test.OK(t, err)
 
-			efi := ExtendedStat(fi)
+			efi := fs.ExtendedStat(fi)
 			test.Assert(t, efi.Size == i || efi.Blocks > 0, "Preallocated size of %v, got size %v block %v", i, efi.Size, efi.Blocks)
 		})
 	}
