+++ internal/fs/file_unix.go
@@ -13,21 +13,6 @@ func fixpath(name string) string {
 	return name
 }
 
-// TempFile creates a temporary file which has already been deleted (on
-// supported platforms)
-func TempFile(dir, prefix string) (f *os.File, err error) {
-	f, err = os.CreateTemp(dir, prefix)
-	if err != nil {
-		return nil, err
-	}
-
-	if err = os.Remove(f.Name()); err != nil {
-		return nil, err
-	}
-
-	return f, nil
-}
-
 // isNotSupported returns true if the error is caused by an unsupported file system feature.
 func isNotSupported(err error) bool {
 	if perr, ok := err.(*os.PathError); ok && perr.Err == syscall.ENOTSUP {
