+++ internal/repository/packer_manager.go
@@ -16,7 +16,7 @@ import (
 	"github.com/restic/restic/internal/restic"
 
 	"github.com/restic/restic/internal/debug"
-	"github.com/restic/restic/internal/fs"
+	"github.com/restic/restic/internal/fileio"
 	"github.com/restic/restic/internal/repository/crypto"
 	"github.com/restic/restic/internal/repository/pack"
 )
@@ -201,7 +201,7 @@ func (r *packerManager) forgetPacker(packer *packer) {
 // created or one is returned that already has some blobs.
 func (r *packerManager) newPacker() (pck *packer, err error) {
 	debug.Log("create new pack")
-	tmpfile, err := fs.TempFile("", "restic-temp-pack-")
+	tmpfile, err := fileio.TempFile("", "restic-temp-pack-")
 	if err != nil {
 		return nil, errors.WithStack(err)
 	}
